+++ src/compiler/ast_native.cc
@@ -306,8 +306,8 @@ class ASTNativeCompiler : public Compiler {
       condition_with_na_check = ExtractCategoricalCondition(t2);
     }
     if (node->children[0]->data_count && node->children[1]->data_count) {
-      const int left_freq = node->children[0]->data_count.value();
-      const int right_freq = node->children[1]->data_count.value();
+      const size_t left_freq = node->children[0]->data_count.value();
+      const size_t right_freq = node->children[1]->data_count.value();
       condition_with_na_check
         = fmt::format(" {keyword}( {condition} ) ",
             "keyword"_a = ((left_freq > right_freq) ? "LIKELY" : "UNLIKELY"),
@@ -333,7 +333,7 @@ class ASTNativeCompiler : public Compiler {
   template <typename ThresholdType, typename LeafOutputType>
   void HandleTUNode(const TranslationUnitNode* node,
                     const std::string& dest,
-                    int indent) {
+                    size_t indent) {
     const int unit_id = node->unit_id;
     const std::string new_file = fmt::format("tu{}.c", unit_id);
     const std::string leaf_output_type
